fix: badges (#1563)
