+++ packages/elements-core/src/components/Docs/HttpService/HttpService.tsx
@@ -1,21 +1,15 @@
-import { Badge, Box, Flex, Heading, VStack } from '@stoplight/mosaic';
+import { Box, Flex, Heading, VStack } from '@stoplight/mosaic';
 import { withErrorBoundary } from '@stoplight/react-error-boundary';
 import { IHttpService } from '@stoplight/types';
 import * as React from 'react';
 
-import { badgeDefaultColor } from '../../../constants';
 import { MockingContext } from '../../../containers/MockingProvider';
 import { MarkdownViewer } from '../../MarkdownViewer';
 import { PoweredByLink } from '../../PoweredByLink';
 import { DocsComponentProps } from '..';
+import { VersionBadge } from '../HttpOperation/Badges';
 import { SecuritySchemes } from './SecuritySchemes';
 import { ServerInfo } from './ServerInfo';
-
-const enhanceVersionString = (version: string): string => {
-  if (version[0] === 'v') return version;
-
-  return `v${version}`;
-};
 
 export type HttpServiceProps = DocsComponentProps<Partial<IHttpService>>;
 
@@ -47,7 +41,7 @@ const HttpServiceComponent = React.memo<HttpServiceProps>(({ className, data, lo
 
       {data.version && (
         <Box mt={3}>
-          <Badge style={{ backgroundColor: badgeDefaultColor }}>{enhanceVersionString(data.version)}</Badge>
+          <VersionBadge value={data.version} />
         </Box>
       )}
